+++ docker/resource_docker_container.go
@@ -24,6 +24,12 @@ func resourceDockerContainer() *schema.Resource {
 				Optional: true,
 			},
 
+			"start": &schema.Schema{
+				Type:     schema.TypeBool,
+				Default:  true,
+				Optional: true,
+			},
+
 			// Indicates whether the container must be running.
 			//
 			// An assumption is made that configured containers