+++ docker/resource_docker_container_test.go
@@ -801,6 +801,22 @@ func TestAccDockerContainer_healthcheck(t *testing.T) {
 	})
 }
 
+func TestAccDockerContainer_nostart(t *testing.T) {
+	var c types.ContainerJSON
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDockerContainerNoStartConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccContainerNotRunning("docker_container.foo", &c),
+				),
+			},
+		},
+	})
+}
+
 func testAccContainerRunning(n string, container *types.ContainerJSON) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
@@ -839,33 +855,19 @@ func testAccContainerNotRunning(n string, container *types.ContainerJSON) resour
 		if !ok {
 			return fmt.Errorf("Not found: %s", n)
 		}
-
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("No ID is set")
 		}
-
 		client := testAccProvider.Meta().(*ProviderConfig).DockerClient
-		containers, err := client.ContainerList(context.Background(), types.ContainerListOptions{
-			All: true,
-		})
+		containers, err := client.ContainerList(context.Background(), types.ContainerListOptions{})
 		if err != nil {
 			return err
 		}
-
 		for _, c := range containers {
 			if c.ID == rs.Primary.ID {
-				inspected, err := client.ContainerInspect(context.Background(), c.ID)
-				if err != nil {
-					return fmt.Errorf("Container could not be inspected: %s", err)
-				}
-				*container = inspected
-
-				if container.State.Running {
-					return fmt.Errorf("Container is running: %s", rs.Primary.ID)
-				}
+				return fmt.Errorf("Container found: %s", rs.Primary.ID)
 			}
 		}
-
 		return nil
 	}
 }
@@ -1288,5 +1290,18 @@ resource "docker_container" "foo" {
     start_period = "15s"
     retries      = 10
   }
+}
+`
+const testAccDockerContainerNoStartConfig = `
+resource "docker_image" "foo" {
+  name         = "nginx:latest"
+  keep_locally = true
+}
+
+resource "docker_container" "foo" {
+  name     = "tf-test"
+  image    = "nginx:latest"
+  start    = false
+  must_run = false
 }
 `