+++ CHANGELOG.md
@@ -4,6 +4,7 @@ IMPROVEMENTS
 * Adds labels for `network`, `volume` and `secret` to support docker stacks. [[#92](https://github.com/terraform-providers/terraform-provider-docker/pull/92)] 
 * Adds `rm` and `attach` options to execute short-lived containers [GH-43] and [[#106](https://github.com/terraform-providers/terraform-provider-docker/pull/106)]
 * Adds container healthcheck[[#93](https://github.com/terraform-providers/terraform-provider-docker/pull/93)]
+* Adds the docker container start flag [GH-62] and [[#94](https://github.com/terraform-providers/terraform-provider-docker/pull/94)]
 
 BUG FIXES
 * Fixes that new network were appended to the default bridge [GH-10]